+++ simulation/traffic_simulator/include/traffic_simulator/entity/ego_entity.hpp
@@ -76,8 +76,7 @@ public:
     const std::string & name,             //
     const Configuration & configuration,  //
     const double step_time,               //
-    const openscenario_msgs::msg::VehicleParameters & parameters,
-    const openscenario_msgs::msg::EntityStatus & status);
+    const openscenario_msgs::msg::VehicleParameters & parameters);
 
   explicit EgoEntity(EgoEntity &&) = delete;
 
@@ -118,9 +117,13 @@ public:
 
   void requestLaneChange(const std::int64_t) override;
 
-  void setStatus(const openscenario_msgs::msg::EntityStatus &) override;
+  auto setDriverModel(const openscenario_msgs::msg::DriverModel &) -> void override;
+
+  auto setStatus(const openscenario_msgs::msg::EntityStatus & status) -> bool override;
 
   void setTargetSpeed(double, bool) override;
+
+  auto setUpperBoundSpeed(double) -> void override;
 };
 }  // namespace entity
 }  // namespace traffic_simulator