+++ simulation/traffic_simulator/src/entity/entity_base.cpp
@@ -25,21 +25,15 @@ namespace traffic_simulator
 {
 namespace entity
 {
-EntityBase::EntityBase(
-  const std::string & type, const std::string & name,
-  const openscenario_msgs::msg::EntityStatus & status)
-: type(type),
-  name(name),
-  status_(status),
-  verbose_(true),
-  visibility_(true),
-  stand_still_duration_(0)
+EntityBase::EntityBase(const std::string & type, const std::string & name)
+: type(type), name(name), status_(boost::none), verbose_(true), visibility_(true)
 {
+  status_ = boost::none;
 }
 
 void EntityBase::onUpdate(double, double) { status_before_update_ = status_; }
 
-double EntityBase::getStandStillDuration() const { return stand_still_duration_; }
+boost::optional<double> EntityBase::getStandStillDuration() const { return stand_still_duration_; }
 
 const autoware_vehicle_msgs::msg::VehicleCommand EntityBase::getVehicleCommand()
 {
@@ -48,47 +42,71 @@ const autoware_vehicle_msgs::msg::VehicleCommand EntityBase::getVehicleCommand()
 
 void EntityBase::updateStandStillDuration(const double step_time)
 {
-  if (std::fabs(status_.action_status.twist.linear.x) <= std::numeric_limits<double>::epsilon()) {
-    stand_still_duration_ = step_time + stand_still_duration_;
+  if (!status_) {
+    stand_still_duration_ = boost::none;
   } else {
-    stand_still_duration_ = 0;
+    if (!stand_still_duration_) {
+      stand_still_duration_ = 0;
+    }
+    if (
+      std::fabs(status_->action_status.twist.linear.x) <= std::numeric_limits<double>::epsilon()) {
+      stand_still_duration_ = step_time + stand_still_duration_.get();
+    } else {
+      stand_still_duration_ = 0;
+    }
   }
 }
 
 void EntityBase::updateEntityStatusTimestamp(const double current_time)
 {
-  status_.time = current_time;
+  if (status_) {
+    status_->time = current_time;
+  }
 }
 
 void EntityBase::setOtherStatus(
   const std::unordered_map<std::string, openscenario_msgs::msg::EntityStatus> & status)
 {
   other_status_.clear();
-  for (const auto & each : status) {
-    if (each.first != name) {
-      const auto p0 = each.second.pose.position;
-      const auto p1 = status_.pose.position;
-      double distance =
-        std::sqrt(std::pow(p0.x - p1.x, 2) + std::pow(p0.y - p1.y, 2) + std::pow(p0.z - p1.z, 2));
-      if (distance < 30) {
-        other_status_.insert(each);
+  if (status_) {
+    for (const auto & each : status) {
+      if (each.first != name) {
+        const auto p0 = each.second.pose.position;
+        const auto p1 = status_.get().pose.position;
+        double distance =
+          std::sqrt(std::pow(p0.x - p1.x, 2) + std::pow(p0.y - p1.y, 2) + std::pow(p0.z - p1.z, 2));
+        if (distance < 30) {
+          other_status_.insert(each);
+        }
       }
     }
   }
 }
 
-const openscenario_msgs::msg::EntityStatus EntityBase::getStatus() const { return this->status_; }
+const openscenario_msgs::msg::EntityStatus EntityBase::getStatus() const
+{
+  if (!status_) {
+    THROW_SEMANTIC_ERROR("status is not set");
+  } else {
+    return this->status_.get();
+  }
+}
 
-void EntityBase::setStatus(const openscenario_msgs::msg::EntityStatus & status)
+bool EntityBase::setStatus(const openscenario_msgs::msg::EntityStatus & status)
 {
   status_ = status;
-  status_.name = name;
+  status_->name = name;
+  return true;
 }
 
 void EntityBase::stopAtEndOfRoad()
 {
-  status_.action_status.twist = geometry_msgs::msg::Twist();
-  status_.action_status.accel = geometry_msgs::msg::Accel();
+  if (!status_) {
+    THROW_SEMANTIC_ERROR("status is not set");
+  } else {
+    status_.get().action_status.twist = geometry_msgs::msg::Twist();
+    status_.get().action_status.accel = geometry_msgs::msg::Accel();
+  }
 }
 }  // namespace entity
 }  // namespace traffic_simulator